Adds Challonge basic integration
