+++ vite.config.ts
@@ -13,11 +13,9 @@ export default defineConfig({
     },
 	server: {
 		proxy: {
-			// // string shorthand: /foo -> http://localhost:4567/foo
-			// '/foo': 'http://localhost:5173',
-			// with options
+
 			'/api': {
-			  target: 'http://localhost:3000',
+			  target: 'http://localhost:3000/api',
 			  changeOrigin: true,
 			  rewrite: path => path.replace(/^\/api/, '')
 			}
